# docs/swagger.yaml
definitions:
  collegemanagementsystem.Borrow_records:
    properties:
      book_id:
        type: integer
      borrow_date:
        type: string
      borrow_id:
        type: integer
      return_date:
        type: string
      user_id:
        type: integer
      user_type:
        type: string
    type: object
  collegemanagementsystem.BorrowInfo:
    properties:
      book_id:
        type: integer
      book_type:
        type: string
      borrow_date:
        type: string
      borrow_id:
        type: integer
      return_date:
        type: string
      user_id:
        type: integer
      user_type:
        type: string
    type: object
...
  collegemanagementsystem.Library:
    properties:
      author:
        type: string
      available_copies:
        type: integer
      book_id:
        type: integer
      book_name:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/Libraries/{id}:
    get:
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/collegemanagementsystem.Library'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get library by ID
      tags:
      - Libraries
    put:
      consumes:
      - application/json
      parameters:
      - description: Updated Library
        in: body
        name: library
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Library'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/collegemanagementsystem.Library'
      security:
      - BearerAuth: []
      summary: Update library
      tags:
      - Libraries
  /api/borrow:
    get:
      description: Retrieve complete borrowing history with book details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/collegemanagementsystem.BorrowInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all borrow records
      tags:
      - Borrow
    post:
      consumes:
      - application/json
      parameters:
      - description: Borrow Record
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Borrow_records'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Borrow book
      tags:
      - Borrow
...
  /api/libraries:
    post:
      consumes:
      - application/json
      parameters:
      - description: Library Book
        in: body
        name: library
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Library'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/collegemanagementsystem.Library'
      security:
      - BearerAuth: []
      summary: Add library book
      tags:
      - Library
  /api/libraries/{id}:
    delete:
      parameters:
      - description: Library ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: library deleted
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete library
      tags:
      - Libraries
  /api/return:
    post:
      consumes:
      - application/json
      parameters:
      - description: Return Record
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Borrow_records'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Return book
      tags:
      - Borrow
...
swagger: "2.0"